Add files via upload
Bounce is a personal game project that I created as part of my journey to learn and explore game development using the Raylib library. The project is intended to help me gain hands-on experience with game mechanics, graphics rendering, and basic game logic. It's a small-scale, experimental game where I applied the core features of Raylib to create a playable 2D game